+++ filler.go
@@ -38,7 +38,10 @@ func (r *routeFiller) fill(ctx reflect.Value) error {
 			param := r.reqURL[baseRouteLen+idx]
 			index := r.routeParams[idx]
 			field := ctx.Elem().Field(index)
-			v := parseParameter(field, param)
+			v, err := parseParameter(field.Type(), param)
+			if err != nil {
+				return err
+			}
 			field.Set(v)
 		}
 	}
@@ -53,8 +56,11 @@ func (q *queryFiller) fill(ctx reflect.Value) error {
 	for k, idx := range q.queryParams {
 		field := ctx.Elem().Field(idx)
 		if v, ok := q.query[k]; ok {
-			p := v[0]
-			value := parseParameter(field, p)
+			param := v[0]
+			value, err := parseParameter(field.Type(), param)
+			if err != nil {
+				return err
+			}
 			field.Set(value)
 		}
 	}
@@ -74,7 +80,10 @@ func (f *formFiller) fill(ctx reflect.Value) error {
 		if v, ok := f.form[k]; ok {
 			field := ctx.Elem().Field(idx)
 			p := v[0]
-			value := parseParameter(field, p)
+			value, err := parseParameter(field.Type(), p)
+			if err != nil {
+				return err
+			}
 			field.Set(value)
 		}
 	}