+++ parse_test.go
@@ -1,10 +1,10 @@
 package rocket
 
 import (
-	"github.com/dannypsnl/assert"
+	"reflect"
 	"testing"
 
-	"reflect"
+	"github.com/dannypsnl/assert"
 )
 
 func TestParseParameter(t *testing.T) {
@@ -88,8 +88,8 @@ func TestParseParameter(t *testing.T) {
 
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			rv := reflect.Zero(reflect.TypeOf(testCase.expectedValue))
-			v := parseParameter(rv, testCase.paramStr)
+			v, err := parseParameter(reflect.TypeOf(testCase.expectedValue), testCase.paramStr)
+			assert.Eq(err, nil)
 			assert.Eq(v.Interface(), testCase.expectedValue)
 		})
 	}