+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - feat: `func File(filepath string) *Response` at package `response`, use to create a file response with default content-type
 - (#126) fix: duplicate path would panic now
 - (#125) feature: support auto implements OPTIONS method
+- (#86) feature: optional field
 
 ## v0.12.9